Fix warning message: "Interpolation-only expressions are deprecated"
Terraform 0.11 and earlier required all non-constant expressions to be
provided via interpolation syntax, but this pattern is now deprecated. To
silence this warning, remove the "${ sequence from the start and the }"
sequence from the end of this expression, leaving just the inner expression.

Since the required terraform version is ">= 0.12" for this module removed interpolation-only expressions from outputs.tf file